feat: move ward-registered toast to global layout
